+++ src/main/java/io/electrum/billpay/api/AccountLookupsResource.java
@@ -10,6 +10,7 @@ import javax.ws.rs.container.AsyncResponse;
 import javax.ws.rs.container.Suspended;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Request;
 import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
@@ -28,19 +29,52 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.Authorization;
 
-@Path("/accountLookups")
-@Consumes({ "application/json" })
-@Produces({ "application/json" })
+@Path(AccountLookupsResource.PATH)
+@Consumes({ MediaType.APPLICATION_JSON })
+@Produces({ MediaType.APPLICATION_JSON })
 @Api(description = "the accountLookups API", authorizations = { @Authorization("httpBasic") })
 public abstract class AccountLookupsResource {
 
    protected abstract IAccountLookupsResource getResourceImplementation();
 
+   public static final String PATH = BillpayApi.API_BASE_PATH + "/accountLookups";
+
+   public class RequestAccountInfo {
+      public static final String OPERATION = "requestAccountInfo";
+      public static final int SUCCESS = 200;
+      public static final String RELATIVE_PATH = "/" + "{" + PathParameters.REQUEST_ID + "}";
+      public static final String FULL_PATH = AccountLookupsResource.PATH + RELATIVE_PATH;
+
+      public class PathParameters {
+         public static final String REQUEST_ID = "requestId";
+      }
+   }
+
+   public class RequestTrafficFineInfo {
+      public static final String OPERATION = "requestTrafficFineInfo";
+      public static final int SUCCESS = 200;
+      public static final String RELATIVE_PATH = "/traffic/" + "{" + PathParameters.REQUEST_ID + "}";
+      public static final String FULL_PATH = AccountLookupsResource.PATH + RELATIVE_PATH;
+
+      public class PathParameters {
+         public static final String REQUEST_ID = "requestId";
+      }
+   }
+
+   public class RequestPolicyInfo {
+      public static final String OPERATION = "requestPolicyInfo";
+      public static final int SUCCESS = 200;
+      public static final String RELATIVE_PATH = "/policy/" + "{" + PathParameters.REQUEST_ID + "}";
+      public static final String FULL_PATH = AccountLookupsResource.PATH + RELATIVE_PATH;
+
+      public class PathParameters {
+         public static final String REQUEST_ID = "requestId";
+      }
+   }
+
    @POST
-   @Path("/{requestId}")
-   @Consumes({ "application/json" })
-   @Produces({ "application/json" })
-   @ApiOperation(value = "Basic account information request", notes = "Requests the current status of the account identified by the supplied account reference. Not all services support account lookups. In the case where this function is not supported for the requested service, a 501 HTTP status code may be returned")
+   @Path(RequestAccountInfo.RELATIVE_PATH)
+   @ApiOperation(value = "Basic account information request", nickname = RequestAccountInfo.OPERATION, notes = "Requests the current status of the account identified by the supplied account reference. Not all services support account lookups. In the case where this function is not supported for the requested service, a 501 HTTP status code may be returned")
    @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = AccountLookupResponse.class),
          @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
          @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
@@ -48,7 +82,7 @@ public abstract class AccountLookupsResource {
          @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
          @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
    public void requestAccountInfo(
-         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("requestId") String requestId,
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam(RequestAccountInfo.PathParameters.REQUEST_ID) String requestId,
          @ApiParam(value = "An account lookup request", required = true) AccountLookupRequest body,
          @Context SecurityContext securityContext,
          @Suspended AsyncResponse asyncResponse,
@@ -69,10 +103,8 @@ public abstract class AccountLookupsResource {
    }
 
    @POST
-   @Path("/traffic/{requestId}")
-   @Consumes({ "application/json" })
-   @Produces({ "application/json" })
-   @ApiOperation(value = "Basic traffic fine info request", notes = "Requests information related to a specific traffic fine as identified by the noticeNumber. In the case where this function is not supported by the relevant service provider, a 501 HTTP status code may be returned")
+   @Path(RequestTrafficFineInfo.RELATIVE_PATH)
+   @ApiOperation(value = "Basic traffic fine info request", nickname = RequestTrafficFineInfo.OPERATION, notes = "Requests information related to a specific traffic fine as identified by the noticeNumber. In the case where this function is not supported by the relevant service provider, a 501 HTTP status code may be returned")
    @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = TrafficFineLookupResponse.class),
          @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
          @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
@@ -80,7 +112,7 @@ public abstract class AccountLookupsResource {
          @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
          @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
    public void requestTrafficFineInfo(
-         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("requestId") String requestId,
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam(RequestTrafficFineInfo.PathParameters.REQUEST_ID) String requestId,
          @ApiParam(value = "A traffic fine lookup request", required = true) TrafficFineLookupRequest body,
          @Context SecurityContext securityContext,
          @Suspended AsyncResponse asyncResponse,
@@ -101,10 +133,8 @@ public abstract class AccountLookupsResource {
    }
 
    @POST
-   @Path("/policy/{requestId}")
-   @Consumes({ "application/json" })
-   @Produces({ "application/json" })
-   @ApiOperation(value = "Basic policy info request", notes = "Requests the current status of the policy identified by the supplied policy reference. Not all service providers support policy lookups. In the case where this function is not supported by the requested service provider, a 501 HTTP status code may be returned")
+   @Path(RequestPolicyInfo.RELATIVE_PATH)
+   @ApiOperation(value = "Basic policy info request", nickname = RequestPolicyInfo.OPERATION, notes = "Requests the current status of the policy identified by the supplied policy reference. Not all service providers support policy lookups. In the case where this function is not supported by the requested service provider, a 501 HTTP status code may be returned")
    @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = PolicyLookupResponse.class),
          @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
          @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
@@ -112,7 +142,7 @@ public abstract class AccountLookupsResource {
          @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
          @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
    public void requestPolicyInfo(
-         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("requestId") String requestId,
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam(RequestPolicyInfo.PathParameters.REQUEST_ID) String requestId,
          @ApiParam(value = "A policy lookup request", required = true) PolicyLookupRequest body,
          @Context SecurityContext securityContext,
          @Suspended AsyncResponse asyncResponse,
@@ -131,4 +161,16 @@ public abstract class AccountLookupsResource {
             httpHeaders,
             uriInfo);
    }
+
+   /**
+    * Use <code>RequestAccountInfo.OPERATION, RequestTrafficFine.OPERATION, RequestPolicyInfo.OPERATION </code>,
+    * instead.
+    */
+   @Deprecated
+   public class Operations {
+      public static final String REQUEST_POLICY_INFO = "requestPolicyInfo";
+      public static final String REQUEST_TRAFFIC_FINE_INFO = "requestTrafficFineInfo";
+      public static final String REQUEST_ACCOUNT_INFO = "requestAccountInfo";
+   }
+
 }