+++ src/main/java/io/electrum/billpay/api/BillpayApi.java
@@ -1,0 +1,6 @@
+package io.electrum.billpay.api;
+
+public class BillpayApi {
+   public static final String VERSION = "";
+   public static final String API_BASE_PATH = "" + VERSION;
+}