+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,5 +1,14 @@
 This page describes changes to the Billpay Service Interface implemented across different releases of the interface.
 
+## v4.6.0
+
+Released 31 August 2018
+
+- Update Service Interface Base version to v3.14.0
+- Update to support Circle 2.0
+- Make paths constants 
+- Update Hugo build to use non-deprecated now function.  
+
 ## v4.5.0
 
 Released 8 November 2017